// api/src/modules/language/constants/enums/errors.enum.ts
export enum ErrorMessage {
  LANGUAGE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [ngôn ngữ].',

  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',

  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_LANGUAGE_ERROR = 'Lưu thông tin [ngôn ngữ] thất bại.',
  LANGUAGE_NOT_FOUND_ERROR = '[Ngôn ngữ] không tồn tại (id: %s)',
  LANGUAGE_HAS_EXIST_ERROR = '[Ngôn ngữ] đã tồn tại (name: %s)',
}